Changed the default 'orientation' of basis matrices.
All the programs that are a part of lattice-tools now treats the rows of
the basis matrices as the basis vectors of the lattice. In practice this
is implemented as a switch in the default trnaspose option. The internal
representation of the lattices still uses columns as the basis (with
lat-gen being the exception).